Add pressure post-processing and u_pres sign change
To plot pressure contour one must first convert the
total pressure (obtained from the sim.) to static
pressure by: P_total = P + 0.5 * |U|^2. Note that U
obtained from sim. is not velocity but rather
the U*h. Moreover, a reference pressure was defined
which was taken to be the centre of the mesh.

Sign of u_pres was changed due to wrong implemen-
tation w.r.t solver written. That is, u_pres
are the known values on the LHS of the equation.